update to new proc macro semantics
